requirements.txt
cycler==0.10.0
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.5
opencv-python==4.5.1.48
pandas==1.2.0
Pillow==8.1.0
pyparsing==2.4.7
pyrealsense2==2.41.0.2666
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
